pyproject.toml
[tool.poetry]
name = "monitoring"
version = "0.1.0"
description = ""
authors = ["business.siegel@protonmail.com"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
attrs = "^23.1.0"
pydantic = "1.10.7"
amqtt  = "^0.10.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
